refactor(utils): groupement des *CoursName.js dans un même dossier utils
