+++ utils/cours/ColorCoursName.js
