+++ utils/cours/EmojiCoursName.js
