+++ utils/cours/FormatCoursName.js
