+++ views/Devoirs/CreateHomeworkScreen.tsx
@@ -12,10 +12,10 @@ import DateTimePicker from '@react-native-community/datetimepicker';
 import NativeList from '../../components/NativeList';
 import NativeItem from '../../components/NativeItem';
 import NativeText from '../../components/NativeText';
-import { getSavedCourseColor } from '../../utils/ColorCoursName';
+import { getSavedCourseColor } from '../../utils/cours/ColorCoursName';
 import { useAppContext } from '../../utils/AppContext';
 import PapillonLoading from '../../components/PapillonLoading';
-import formatCoursName from '../../utils/FormatCoursName';
+import formatCoursName from '../../utils/cours/FormatCoursName';
 import AlertBottomSheet from '../../interface/AlertBottomSheet';
 import { AlertTriangle } from 'lucide-react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
